Minor cleanup Renamed retirementCalculator to mainPage.js
