+++ .env.dev
@@ -1,6 +1,6 @@
 REACT_APP_ENV = "dev"
 REACT_APP_SERVICE_URL = "https://localhost:5001/api/Retirement/Report"
 REACT_APP_SPENDING = 20000
-REACT_APP_SALARY = "50000"
-REACT_APP_SAVINGS ="50000"
-REACT_APP_PENSION ="50000"
+REACT_APP_SALARY =  50000
+REACT_APP_SAVINGS = 50000
+REACT_APP_PENSION = 50000
